requirements.txt
pandas==0.24.2
ta==0.6.1
numpy==1.19.4
matplotlib==3.1.1
pandas_datareader==0.9.0
spacy==2.3.2
plac==0.9.6
TA_Lib==0.4.19
yfinance==0.1.55
spacy_streamlit==0.1.0
requests==2.24.0
Pillow==8.0.1
scikit_learn==0.23.2
talib==0.1.1
